requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyautogui==0.9.54
keyboard==0.13.5
Pillow==10.0.0
customtkinter==5.2.0
psutil==5.9.5
pywin32==306
mysql-connector-python==8.2.0 